+++ Runtime/Scripts/RTCRtpReceiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc1986e3122cd3547b6cd7678ef43b6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
