+++ Runtime/Scripts/RTCRtpSender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40a73bff38102dd4d9b3925c32f0cc86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
