+++ Runtime/Scripts/WebRTC.cs
@@ -49,6 +49,7 @@ namespace Unity.WebRTC
             protocol = "";
         }
     }
+
     public class RTCRtpReceiver
     {
         internal IntPtr self;
@@ -63,50 +64,6 @@ namespace Unity.WebRTC
         public RTCStatsReportAsyncOperation GetStats()
         {
             return peer.GetStats(this);
-        }
-    }
-
-    public class RTCRtpSender
-    {
-        internal IntPtr self;
-        private RTCPeerConnection peer;
-
-        internal RTCRtpSender(IntPtr ptr, RTCPeerConnection peer)
-        {
-            self = ptr;
-            this.peer = peer;
-        }
-
-        public RTCStatsReportAsyncOperation GetStats()
-        {
-            return peer.GetStats(this);
-        }
-
-        public MediaStreamTrack Track
-        {
-            get
-            {
-                IntPtr ptr = NativeMethods.SenderGetTrack(self);
-                return WebRTC.FindOrCreate<MediaStreamTrack>(ptr, _ptr => new MediaStreamTrack(_ptr));
-            }
-        }
-
-        public RTCRtpSendParameters GetParameters()
-        {
-            NativeMethods.SenderGetParameters(self, out var ptr);
-            RTCRtpSendParametersInternal parametersInternal = Marshal.PtrToStructure<RTCRtpSendParametersInternal>(ptr);
-            RTCRtpSendParameters parameters = new RTCRtpSendParameters(parametersInternal);
-            Marshal.FreeHGlobal(ptr);
-            return parameters;
-        }
-
-        public RTCErrorType SetParameters(RTCRtpSendParameters parameters)
-        {
-            IntPtr ptr = parameters.CreatePtr();
-            RTCErrorType error = NativeMethods.SenderSetParameters(self, ptr);
-            RTCRtpSendParameters.DeletePtr(ptr);
-
-            return error;
         }
     }
 
@@ -195,7 +152,7 @@ namespace Unity.WebRTC
         BundlePolicyBalanced,
         BundlePolicyMaxBundle,
         BundlePolicyMaxCompat
-    };
+    }
 
     public struct RTCSessionDescription
     {
